engineering-pdf-extractor/requirements.txt
pdfplumber
pymupdf
camelot-py
pandas
ghostscript
